src/requirements.txt
cycler==0.10.0
kiwisolver==1.3.1
matplotlib==3.2.2
numpy==1.19.5
Pillow==7.1.2
pyparsing==2.4.7
python-dateutil==2.8.2
six==1.16.0
